Capstone_Project-CS[01]/Codebase/requirements.txt
# Core LLM and AI
langchain>=1.2.13
langchain-community>=0.4.1
langchain-openai>=1.1.11
openai>=2.26.0

# Environment and configuration
python-dotenv>=1.0.0

# PDF and document processing
pdfplumber>=0.10.3
python-docx>=1.1.0
fpdf2>=2.7.6
reportlab>=4.0.7

# Note: Semantic similarity features (matcher.py, ats_checker.py)
# Compatible with openai>=2.26.0
